pull

+++ FShop/src/main/java/Controllers/BuyProductsServlet.java
@@ -257,14 +257,15 @@ public class BuyProductsServlet extends HttpServlet {
                     }
                     if (session.getAttribute("customerVoucherUsing") != null) {
                         CustomerVoucherDAO cv = new CustomerVoucherDAO();
+                        System.out.println(customerVoucherUsing.getQuantity());
                         if (customerVoucherUsing.getQuantity() == 1) {
                             cv.deleteVoucher(cus.getId(), customerVoucherUsing.getVoucherID());
                         } else if (customerVoucherUsing.getQuantity() > 1) {
                             cv.subtractQuantityOfVoucher(cus.getId(), customerVoucherUsing.getVoucherID());
                         }
+                        cv.increaseVoucher(customerVoucherUsing.getVoucherID());
                     }
                 }
-
                 session.setAttribute("orderStatus", "success");
                 session.setAttribute("numOfProCartOfCus", ca.getNumberOfProduct(cus.getId()));
                 //gui mail xac nhan don hang thanh cong
@@ -334,18 +335,18 @@ public class BuyProductsServlet extends HttpServlet {
             Timestamp endDate = Timestamp.valueOf(endDateString);
             LocalDateTime currentDate = LocalDateTime.now();
             boolean isDeleted = false;
-            if (expirationDate.toLocalDateTime().isBefore(currentDate)) {
+            if (expirationDate.toLocalDateTime().isBefore(currentDate) && customerVoucher.getExpirationDate() != null) {
                 System.out.println("Voucher Het Han");
                 c.deleteVoucher(customerId, customerVoucher.getVoucherID());
                 isDeleted = true;
             }
-            if ((customerVoucher.getUsedCount() == customerVoucher.getMaxUsedCount()) && isDeleted == false) {
+            if (((customerVoucher.getUsedCount() == customerVoucher.getMaxUsedCount()) && isDeleted == false) && customerVoucher.getMaxUsedCount() != 0) {
                 System.out.println("Voucher Het Luot sd");
                 c.deleteVoucher(customerId, customerVoucher.getVoucherID());
                 isDeleted = true;
             }
 
-            if (endDate.toLocalDateTime().isBefore(currentDate) && isDeleted == false) {
+            if (endDate.toLocalDateTime().isBefore(currentDate) && isDeleted == false && customerVoucher.getEndDate() != null) {
                 System.out.println("Voucher Het End date");
                 c.deleteVoucher(customerId, customerVoucher.getVoucherID());
                 isDeleted = true;
